# docs/swagger.json
{"openapi":"3.0.0","paths":{"/v1/app/health":{"get":{"operationId":"AppController_health","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["app"]}},"/v1/auth/sign-up":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpInputDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpResponseDto"}}}}},"tags":["auth"]}},"/v1/auth/sign-in":{"post":{"operationId":"AuthController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInInputDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInResponseDto"}}}}},"tags":["auth"]}},"/v1/users":{"get":{"operationId":"UserController_getUsers","parameters":[{"name":"filter","required":true,"in":"query","explode":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUsersFilter"}}},"schema":{"type":"object"}},{"name":"pageOptions","required":false,"in":"query","explode":false,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PageOptionsInput"}}},"schema":{"type":"object"}}],"responses":{"200":{"description":"Successfully received model list","content":{"application/json":{"schema":{"title":"PaginatedResponseOfUserResponseDto","allOf":[{"$ref":"#/components/schemas/PaginationResponse"},{"properties":{"data":{"type":"array","items":{"$ref":"#/components/schemas/UserResponseDto"}}}}]}}}}},"tags":["users"],"security":[{"bearer":[]}]},"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserInputDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/v1/users/{id}":{"get":{"operationId":"UserController_getUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"tags":["users"],"security":[{"bearer":[]}]},"patch":{"operationId":"UserController_updateById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserInputDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponseDto"}}}}},"tags":["users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserController_delete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}},"/v1/reddit":{"get":{"operationId":"RedditController_search","parameters":[{"name":"search","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RedditSearchResponse"}}}}},"tags":["reddit"],"security":[{"bearer":[]}]}}},"info":{"title":"Sisal Nest App","description":"Nest Prisma Mysql Rest API Documentation","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignUpInputDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"UserType":{"type":"string","enum":["admin","user"]},"UserRole":{"type":"string","enum":["admin_superAdmin","user_member"]},"SignUpResponseDto":{"type":"object","properties":{"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"email":{"type":"string"},"type":{"$ref":"#/components/schemas/UserType"},"role":{"type":"array","items":{"$ref":"#/components/schemas/UserRole"}}},"required":["id","createdAt","email","type","role"]},"SignInInputDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"SignInResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"expiresAt":{"format":"date-time","type":"string"}},"required":["accessToken","refreshToken","expiresAt"]},"PageOptionsInput":{"type":"object","properties":{"orderDirection":{"type":"string","enum":["asc","desc"],"default":"desc"},"orderBy":{"type":"string","default":"id"},"page":{"type":"number","minimum":1,"default":1},"take":{"type":"number","minimum":1,"maximum":50,"default":10},"search":{"type":"string"}}},"PageMeta":{"type":"object","properties":{"page":{"type":"number"},"take":{"type":"number"},"itemCount":{"type":"number"},"pageCount":{"type":"number"},"hasPreviousPage":{"type":"boolean"},"hasNextPage":{"type":"boolean"}},"required":["page","take","itemCount","pageCount","hasPreviousPage","hasNextPage"]},"PaginationResponse":{"type":"object","properties":{"data":{"type":"array","items":{"type":"object"}},"meta":{"$ref":"#/components/schemas/PageMeta"}},"required":["data","meta"]},"UserResponseDto":{"type":"object","properties":{"id":{"type":"number"},"createdAt":{"format":"date-time","type":"string"},"email":{"type":"string"},"type":{"$ref":"#/components/schemas/UserType"},"isDeleted":{"type":"boolean"},"role":{"type":"array","items":{"$ref":"#/components/schemas/UserRole"}}},"required":["id","createdAt","email","type","isDeleted","role"]},"DateRangeFilter":{"type":"object","properties":{"from":{"format":"date-time","type":"string"},"to":{"format":"date-time","type":"string"}}},"GetUsersFilter":{"type":"object","properties":{"type":{"$ref":"#/components/schemas/UserType"},"role":{"$ref":"#/components/schemas/UserRole"},"createdAtRange":{"$ref":"#/components/schemas/DateRangeFilter"}}},"CreateUserInputDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"},"type":{"$ref":"#/components/schemas/UserType"},"role":{"$ref":"#/components/schemas/UserRole"}},"required":["email","password","type","role"]},"UpdateUserInputDto":{"type":"object","properties":{"email":{"type":"string"},"type":{"$ref":"#/components/schemas/UserType"},"role":{"$ref":"#/components/schemas/UserRole"}}},"SubRedditResponse":{"type":"object","properties":{"title":{"type":"string"},"imageUrl":{"type":"string"}},"required":["title","imageUrl"]},"RedditSearchResponse":{"type":"object","properties":{"subreddits":{"type":"array","items":{"$ref":"#/components/schemas/SubRedditResponse"}},"numberOfImages":{"type":"number"}},"required":["subreddits","numberOfImages"]}}}}